pyproject.toml
[tool.poetry]
name = "shift-ml-python-task"
version = "0.1.0"
description = ""
authors = ["kirillsoulshift <soulshift.twiggy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pytest = "^8.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
